banking_dcpr/requirements.txt
asgiref==3.8.1
Django==5.1.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.11.0
sqlparse==0.5.3
tzdata==2025.1
whitenoise==6.12.0
